# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  core_transport_http_response.ErrorResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short msg
        type: string
    type: object
...
  tasks_service.CreateTaskRequest:
    properties:
      assigned_to:
        type: string
      description:
        maxLength: 2000
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - title
    type: object
  tasks_transport_http.PatchTaskRequest:
    type: object
  tasks_transport_http.TaskDTOResponse:
    properties:
      assigned_to:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      household_id:
        type: string
      id:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  users_transport_http.PatchUserRequest:
    type: object
  users_transport_http.PatchUserResponse:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
  users_transport_http.UserDTOResponse:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
host: 127.0.0.1:5050
info:
  contact: {}
  description: Cohesive Application REST-API scheme
  title: Cohesive-Core
  version: "0.1"
paths:
...
  /households/{id}/tasks:
    get:
      description: Получить список задач дома
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список задач
          schema:
            items:
              $ref: '#/definitions/tasks_transport_http.TaskDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Household not found
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      summary: Список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создать новую задачу в доме
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      - description: CreateTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tasks_service.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Задача создана
          schema:
            $ref: '#/definitions/tasks_transport_http.TaskDTOResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Household not found
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      summary: Создать задачу
      tags:
      - tasks
  /households/{id}/tasks/{task_id}:
    delete:
      description: Удалить задачу по ID
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      - description: ID задачи
        in: path
        name: task_id
        required: true
        type: string
      responses:
        "204":
          description: Задача удалена
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      summary: Удалить задачу
      tags:
      - tasks
    get:
      description: Получить задачу по ID
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      - description: ID задачи
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Задача
          schema:
            $ref: '#/definitions/tasks_transport_http.TaskDTOResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      summary: Получить задачу
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      description: Частично обновить задачу (title, description, status, assigned_to)
      parameters:
      - description: ID дома
        in: path
        name: id
        required: true
        type: string
      - description: ID задачи
        in: path
        name: task_id
        required: true
        type: string
      - description: PatchTask тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tasks_transport_http.PatchTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Задача обновлена
          schema:
            $ref: '#/definitions/tasks_transport_http.TaskDTOResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "409":
          description: Version conflict
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      summary: Обновить задачу
      tags:
      - tasks
...
  /users/me:
    delete:
      description: Удалить текущего пользователя из системы
      responses:
        "204":
          description: Пользователь удалён
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удалить пользователя
      tags:
      - users
    get:
      description: Получить информацию о текущем пользователе
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь получен
          schema:
            $ref: '#/definitions/users_transport_http.UserDTOResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить пользователя
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Частично обновить информацию о текущем пользователе
      parameters:
      - description: PatchMe тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users_transport_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь обновлён
          schema:
            $ref: '#/definitions/users_transport_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "409":
          description: Version conflict
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/core_transport_http_response.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить пользователя
      tags:
      - users
swagger: "2.0"