# internal/features/tasks/transport/http/get_task.go
package tasks_transport_http

import (
	core_errors "cohesive-core/internal/core/errors"
	core_logger "cohesive-core/internal/core/logger"
	core_transport_http_middleware "cohesive-core/internal/core/transport/http/middleware"
	core_transport_http_response "cohesive-core/internal/core/transport/http/response"
	"fmt"
	"net/http"

	"github.com/google/uuid"
)

// GetTask godoc
// @Summary Получить задачу
// @Description Получить задачу по ID
// @Tags tasks
// @Produce json
// @Param id path string true "ID дома"
// @Param task_id path string true "ID задачи"
// @Success 200 {object} tasks_transport_http.TaskDTOResponse "Задача"
// @Failure 400 {object} core_transport_http_response.ErrorResponse "Bad request"
// @Failure 401 {object} core_transport_http_response.ErrorResponse "Unauthorized"
// @Failure 404 {object} core_transport_http_response.ErrorResponse "Task not found"
// @Failure 500 {object} core_transport_http_response.ErrorResponse "Internal server error"
// @Router /households/{id}/tasks/{task_id} [get]
func (h *TasksHTTPHandler) GetTask(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_transport_http_response.NewHTTPResponseHandler(log, rw)

	callerID, ok := core_transport_http_middleware.UserIDFromContext(ctx)
	if !ok {
		responseHandler.ErrorResponse(fmt.Errorf("user id not found in request context"), "internal server errors")

		return
	}

	householdID, err := uuid.Parse(r.PathValue("id"))
	if err != nil {
		responseHandler.ErrorResponse(
			fmt.Errorf("parse household id %q: %v: %w", r.PathValue("id"), err, core_errors.ErrInvalidArgument),
			"invalid household id",
		)

		return
	}

	taskID, err := uuid.Parse(r.PathValue("task_id"))
	if err != nil {
		responseHandler.ErrorResponse(
			fmt.Errorf("parse task id %q: %v: %w", r.PathValue("task_id"), err, core_errors.ErrInvalidArgument),
			"invalid task id",
		)

		return
	}

	task, err := h.tasksService.GetTask(ctx, householdID, taskID, callerID)
	if err != nil {
		responseHandler.ErrorResponse(err, "failed to get task")

		return
	}

	response := taskDTOFromDomain(task)
	responseHandler.JSONResponse(response, http.StatusOK)
}